backend/requirements.txt
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.0
gunicorn
whitenoise
